requirements/base.txt
Django>=5.1,<5.2
dj-database-url>=2.2
psycopg[binary]>=3.1
django-allauth>=65.0
django-htmx>=1.19
django-elasticsearch-dsl>=8.0
djangorestframework>=3.15
celery[redis]>=5.4
django-redis>=5.4
redis>=5.0
elasticsearch-dsl>=8.0
markdown>=3.7
bleach>=6.1
Pillow>=10.0
python-dotenv>=1.0
requests>=2.32
PyJWT>=2.10
cryptography>=45.0
